requirements.txt
pyttsx3
speechrecognition
pyaudio
google-api-python-client
tk
requests
